+++ packages/core/lozenge/__tests__/index.js
@@ -1,36 +1,38 @@
 // @flow
-
-import { mount, shallow } from 'enzyme';
+import { mount } from 'enzyme';
+import renderer from 'react-test-renderer';
 import React from 'react';
 
-import Lozenge, { APPEARANCE_ENUM } from '../src/Lozenge';
+import Lozenge from '../src';
 
 describe('Lozenge', () => {
   describe('isBold property', () => {
     it('should not be the default', () => {
       expect(mount(<Lozenge />).prop('isBold')).toBe(false);
     });
+
     it('should change when toggled', () => {
       expect(mount(<Lozenge isBold />).prop('isBold')).toBe(true);
     });
   });
+
   describe('appearance property', () => {
     it('should be "default" when not set', () => {
       expect(mount(<Lozenge />).prop('appearance')).toBe('default');
     });
-    it('should change when set to an approved value', () => {
-      APPEARANCE_ENUM.values.forEach(value => {
-        expect(mount(<Lozenge appearance={value} />).prop('appearance')).toBe(
-          value,
-        );
-      });
+
+    it('should change when set to an approved value: success', () => {
+      expect(mount(<Lozenge appearance="success" />).prop('appearance')).toBe(
+        'success',
+      );
     });
-    it('should revert to "default" when set to an invalid value', () => {
-      expect(
-        shallow(<Lozenge appearance={('foo': any)} />)
-          .instance()
-          .validAppearance(),
-      ).toBe('default');
+
+    it('should render correctly with text truncated', () => {
+      const wrapper = (
+        <Lozenge appearance="new">Hello, I should truncate at a point.</Lozenge>
+      );
+      const Component = renderer.create(wrapper).toJSON();
+      expect(Component).toMatchSnapshot();
     });
   });
 });