+++ packages/core/navigation-next/src/theme/index.js
@@ -11,3 +11,6 @@ export { styleReducerNoOp } from './utils';
 
 // withTheme
 export { default as withTheme } from './withTheme';
+
+// Types
+export { GlobalTheme } from './types';