+++ packages/editor/conversation/src/containers/Conversation.tsx
@@ -105,6 +105,7 @@ export interface ContainerProps {
   renderEditor?: (Editor: typeof AkEditor, props: EditorProps) => JSX.Element;
   placeholder?: string;
   disableScrollTo?: boolean;
+  allowFeedbackAndHelpButtons?: boolean;
 }
 
 class ConversationContainer extends React.Component<ContainerProps, any> {