+++ packages/editor/editor-core/__tests__/plugins/media/ui/MediaSingleEdit.tsx
@@ -7,12 +7,13 @@ import {
   randomId,
   createEditor,
   bodiedExtension,
+  layoutColumn,
+  layoutSection,
+  p,
 } from '@atlaskit/editor-test-helpers';
-import mediaPlugin from '../../../../src/plugins/media';
 import { stateKey as pluginKey } from '../../../../src/plugins/media/pm-plugins/main';
-import extensionPlugin from '../../../../src/plugins/extension';
-import MediaSingleEdit from '../../../../src/plugins/media/ui/MediaSingleEdit';
 import ToolbarButton from '../../../../src/ui/ToolbarButton';
+import MediaSingleEdit from '../../../../src/plugins/media/ui/MediaSingleEdit';
 
 describe('@atlaskit/editor-core/ui/MediaSingleEdit', () => {
   const testCollectionName = `media-plugin-mock-collection-${randomId()}`;
@@ -20,7 +21,14 @@ describe('@atlaskit/editor-core/ui/MediaSingleEdit', () => {
   const editor = (doc: any) =>
     createEditor({
       doc,
-      editorPlugins: [mediaPlugin({ allowMediaSingle: true }), extensionPlugin],
+      editorProps: {
+        media: {
+          provider: new Promise(() => {}),
+          allowMediaSingle: true,
+        },
+        allowExtension: true,
+        UNSAFE_allowLayouts: true,
+      },
     });
   it('should have layout options if media single not inside bodied extension', () => {
     const { editorView } = editor(
@@ -72,6 +80,44 @@ describe('@atlaskit/editor-core/ui/MediaSingleEdit', () => {
               width: 100,
               height: 100,
             })(),
+          ),
+        ),
+      ),
+    );
+    editorView.focus();
+    const pluginState = pluginKey.getState(editorView.state);
+
+    const mediaSingleEdit = mount(
+      <MediaSingleEdit pluginState={pluginState} />,
+    );
+    expect(editorView.state.selection.$from.node().type.name).toEqual(
+      'mediaSingle',
+    );
+    expect(
+      mediaSingleEdit.findWhere(child => {
+        return child.type() === ToolbarButton && child.prop('disabled');
+      }).length,
+    ).toEqual(3);
+    mediaSingleEdit.unmount();
+  });
+
+  it('should not have layout options if media single inside layoutSection', () => {
+    const { editorView } = editor(
+      doc(
+        layoutSection()(
+          layoutColumn(p('')),
+          layoutColumn(
+            mediaSingle({ layout: 'center' })(
+              media({
+                id: temporaryFileId,
+                __key: temporaryFileId,
+                type: 'file',
+                collection: testCollectionName,
+                __fileMimeType: 'image/png',
+                width: 100,
+                height: 100,
+              })(),
+            ),
           ),
         ),
       ),