+++ packages/editor/editor-core/src/plugins/date/index.tsx
@@ -6,6 +6,12 @@ import { insertDate, selectElement } from './actions';
 import createPlugin, { DateState, pluginKey } from './plugin';
 import DatePicker from './ui/DatePicker';
 
+export type DateType = {
+  year: number;
+  month: number;
+  day?: number;
+};
+
 const datePlugin: EditorPlugin = {
   nodes() {
     return [{ rank: 2400, name: 'date', node: date }];
@@ -32,9 +38,7 @@ const datePlugin: EditorPlugin = {
           dateState.element ? (
             <DatePicker
               element={dateState.element}
-              onSelect={({ iso }) =>
-                insertDate(iso)(editorView.state, dispatch)
-              }
+              onSelect={date => insertDate(date)(editorView.state, dispatch)}
               onClickOutside={() =>
                 selectElement(null)(editorView.state, dispatch)
               }