+++ packages/editor/editor-core/src/styles/index.ts
@@ -2,23 +2,17 @@ import { fontSize } from '@atlaskit/theme';
 import {
   akCodeFontFamily,
   akColorB100,
-  akColorB300,
   akColorB400,
   akColorB50,
-  akColorB75,
   akColorN0,
   akColorN20,
   akColorN30,
   akColorN40,
-  akColorN50,
   akColorN100,
   akColorN500,
   akColorN700,
   akColorN900,
   akZIndexLayer,
-  akColorR50,
-  akColorR300,
-  akColorR75,
 } from '@atlaskit/util-shared-styles';
 
 export const akEditorCodeFontFamily = akCodeFontFamily;
@@ -37,20 +31,10 @@ export const akEditorCodeBlockPadding = '12px';
 export const akEditorCodeInlinePadding = '2px 4px';
 export const akEditorFloatingPanelZIndex = akZIndexLayer;
 export const akEditorMentionSelected = akColorN100;
-export const akEditorTableToolbarSize = 12;
-export const akEditorTableBorder = akColorN50;
-export const akEditorTableToolbar = akColorN20;
-export const akEditorTableFloatingControls = akColorN20;
-export const akEditorTableCellSelected = akColorB75;
-export const akEditorTableToolbarSelected = akColorB100;
-export const akEditorTableBorderSelected = akColorB300;
-export const akEditorTableCellDelete = akColorR50;
-export const akEditorTableBorderDelete = akColorR300;
-export const akEditorTableToolbarDelete = akColorR75;
-export const akEditorTableBorderRadius = '3px';
 export const akEditorRuleBackground = akColorN30;
 export const akEditorRuleBorderRadius = '1px';
 export const defaultEditorFontStyles = `
   font-size: ${fontSize()}px;
   font-weight: normal;
 `;
+export const akEditorToolbarKeylineHeight = 2;