+++ packages/search/global-search/__tests__/SearchResultsUtil.test.tsx
@@ -6,7 +6,15 @@ import {
   searchConfluenceItem,
   AdvancedSearchItemProps,
   searchJiraItem,
+  renderResults,
 } from '../src/components/SearchResultsUtil';
+import {
+  Result,
+  ResultType,
+  AnalyticsType,
+  ResultContentType,
+} from '../src/model/Result';
+import ObjectResult from '../src/components/ObjectResult';
 
 describe('searchPeopleItem', () => {
   function render(partialProps: Partial<AdvancedSearchItemProps>) {
@@ -76,5 +84,33 @@ describe('searchJiraItem', () => {
     expect(wrapper.prop('href')).toEqual(
       '/issues/?jql=text%20~%20%22test%20query%22',
     );
+  });
+
+  describe('renderResults', () => {
+    it('should pass the required properties to the Result component', () => {
+      const result: Result = {
+        resultId: 'resultId',
+        resultType: ResultType.Object,
+        analyticsType: AnalyticsType.RecentConfluence,
+        name: 'name',
+        containerName: 'containerName',
+        href: 'href',
+        avatarUrl: 'avatarUrl',
+        objectKey: 'objectKey',
+        contentType: ResultContentType.Page,
+      };
+
+      const wrapper = shallow(<span>{renderResults([result])}</span>);
+      const resultComponent = wrapper.find(ObjectResult);
+
+      expect(resultComponent.prop('resultId')).toEqual('resultId');
+      expect(resultComponent.prop('type')).toEqual('recent-confluence');
+      expect(resultComponent.prop('name')).toEqual('name');
+      expect(resultComponent.prop('containerName')).toEqual('containerName');
+      expect(resultComponent.prop('href')).toEqual('href');
+      expect(resultComponent.prop('avatarUrl')).toEqual('avatarUrl');
+      expect(resultComponent.prop('objectKey')).toEqual('objectKey');
+      expect(resultComponent.prop('contentType')).toEqual('page');
+    });
   });
 });