+++ packages/search/quick-search/src/components/Results/PersonResult.js
@@ -30,9 +30,19 @@ export default class PersonResult extends PureComponent<Props> {
       ? `${this.props.mentionPrefix}${this.props.mentionName}`
       : null;
 
-  getAvatar = () => (
-    <Avatar presence={this.props.presenceState} src={this.props.avatarUrl} />
-  );
+  getAvatar = () => {
+    if (this.props.avatar) {
+      return this.props.avatar;
+    }
+
+    return (
+      <Avatar
+        presence={this.props.presenceState}
+        size="small"
+        src={this.props.avatarUrl}
+      />
+    );
+  };
 
   render() {
     const { name, presenceMessage, ...resultBaseProps } = this.props;