+++ packages/core/logo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/logo
 
+## 7.0.1
+- [patch] Updated dependencies [d662caa](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/d662caa)
+  - @atlaskit/field-radio-group@2.0.3
+  - @atlaskit/button@7.2.5
+  - @atlaskit/docs@3.0.4
+
 ## 7.0.0
 - [major] Bump to React 16.3. [4251858](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/4251858)